DOCS: documenting and cleaning work from tonight
