+++ .gitsecret/paths/mapping.cfg
@@ -1,1 +1,1 @@
-db/backups/20200408.tar.gz:b4f973f283312971e9850e1b8cc58fcbdec518b76b6f7a52a5a3beb2fb45bdee
+db/backups/20200408.sql.gz:b365bf5fc13dbbe26d93e88f6523451b092999edc2a3f0680482f3a7e7369468
