+++ db/backups/.gitignore
@@ -1,0 +1,3 @@
+*
+!.gitignore
+!*.secret
