--- db/test_edit_distance.py
@@ -1,13 +1,0 @@
-import pandas as pd
-
-from fmp import FMP
-
-fmp_model = FMP("20200303_chihaiti", "hjohnson", "hjohnson")
-
-pk_visit_id = "pk_visit_id"
-fk_clinic_id = "fk_clinic_id"
-
-print(fmp_model.get_tablename_from_fieldname(pk_visit_id))  # expect this to be "visits"
-print(
-    fmp_model.get_tablename_from_fieldname(fk_clinic_id)
-)  # expect this to be "clinics"
